graph_database/chatbot_vectorsearch_filterByKG/requirements.txt
sentence_transformers
elasticsearch
autogen-agentchat
autogen-ext[azure]
aiohttp
azure-core 
langfuse
ragas
fastapi
azure-ai-contentsafety
wikipedia
duckduckgo_search==6.4.2
uvicorn
streamlit
google-genai
llama_cloud_services
llama-index-embeddings-huggingface
langchain-neo4j